refactor(cli): extract hooks.py command module from cli monolith
Third slice of the cli monolith decomposition. Moves the install-hook,
uninstall-hook, init, and setup command implementations together with
their hook-script, hook-message, and stub-message constants from
phi_scan/cli/__init__.py into phi_scan/cli/hooks.py.

Commands are re-registered in __init__.py via
app.command(name)(function) so the user-facing CLI is unchanged. Nine
test monkeypatch targets for _PRE_COMMIT_HOOK_PATH now point at
phi_scan.cli.hooks._PRE_COMMIT_HOOK_PATH (the new home of the constant
that install_hook / uninstall_hook actually read). No behavior change.

cli/__init__.py shrinks 1135 → 1089 lines.
