+++ tests/test_cli.py
@@ -298,7 +298,7 @@ def test_install_hook_creates_hook_file(tmp_path: Path, monkeypatch: pytest.Monk
     hook_path = tmp_path / ".git" / "hooks" / "pre-commit"
     (tmp_path / ".git").mkdir()
     monkeypatch.setattr("phi_scan.cli._shared._GIT_DIR_PATH", tmp_path / ".git")
-    monkeypatch.setattr("phi_scan.cli._PRE_COMMIT_HOOK_PATH", str(hook_path))
+    monkeypatch.setattr("phi_scan.cli.hooks._PRE_COMMIT_HOOK_PATH", str(hook_path))
 
     result = _runner.invoke(app, ["install-hook"])
 
@@ -313,7 +313,7 @@ def test_install_hook_prints_installed_message(
     hook_path = tmp_path / ".git" / "hooks" / "pre-commit"
     (tmp_path / ".git").mkdir()
     monkeypatch.setattr("phi_scan.cli._shared._GIT_DIR_PATH", tmp_path / ".git")
-    monkeypatch.setattr("phi_scan.cli._PRE_COMMIT_HOOK_PATH", str(hook_path))
+    monkeypatch.setattr("phi_scan.cli.hooks._PRE_COMMIT_HOOK_PATH", str(hook_path))
 
     result = _runner.invoke(app, ["install-hook"])
 
@@ -328,7 +328,7 @@ def test_install_hook_does_not_overwrite_existing_hook(
     hook_path.parent.mkdir(parents=True, exist_ok=True)
     hook_path.write_text(_FOREIGN_HOOK_CONTENT)
     monkeypatch.setattr("phi_scan.cli._shared._GIT_DIR_PATH", tmp_path / ".git")
-    monkeypatch.setattr("phi_scan.cli._PRE_COMMIT_HOOK_PATH", str(hook_path))
+    monkeypatch.setattr("phi_scan.cli.hooks._PRE_COMMIT_HOOK_PATH", str(hook_path))
 
     result = _runner.invoke(app, ["install-hook"])
 
@@ -342,7 +342,7 @@ def test_uninstall_hook_removes_our_hook(tmp_path: Path, monkeypatch: pytest.Mon
     hook_path.parent.mkdir(parents=True, exist_ok=True)
     hook_path.write_text(_EXPECTED_HOOK_SCRIPT, encoding="utf-8")
     monkeypatch.setattr("phi_scan.cli._shared._GIT_DIR_PATH", tmp_path / ".git")
-    monkeypatch.setattr("phi_scan.cli._PRE_COMMIT_HOOK_PATH", str(hook_path))
+    monkeypatch.setattr("phi_scan.cli.hooks._PRE_COMMIT_HOOK_PATH", str(hook_path))
 
     result = _runner.invoke(app, ["uninstall-hook"])
 
@@ -358,7 +358,7 @@ def test_uninstall_hook_prints_removed_message(
     hook_path.parent.mkdir(parents=True, exist_ok=True)
     hook_path.write_text(_EXPECTED_HOOK_SCRIPT, encoding="utf-8")
     monkeypatch.setattr("phi_scan.cli._shared._GIT_DIR_PATH", tmp_path / ".git")
-    monkeypatch.setattr("phi_scan.cli._PRE_COMMIT_HOOK_PATH", str(hook_path))
+    monkeypatch.setattr("phi_scan.cli.hooks._PRE_COMMIT_HOOK_PATH", str(hook_path))
 
     result = _runner.invoke(app, ["uninstall-hook"])
 
@@ -371,7 +371,7 @@ def test_uninstall_hook_prints_not_found_when_absent(
     hook_path = tmp_path / ".git" / "hooks" / "pre-commit"
     (tmp_path / ".git").mkdir()
     monkeypatch.setattr("phi_scan.cli._shared._GIT_DIR_PATH", tmp_path / ".git")
-    monkeypatch.setattr("phi_scan.cli._PRE_COMMIT_HOOK_PATH", str(hook_path))
+    monkeypatch.setattr("phi_scan.cli.hooks._PRE_COMMIT_HOOK_PATH", str(hook_path))
 
     result = _runner.invoke(app, ["uninstall-hook"])
 
@@ -387,7 +387,7 @@ def test_uninstall_hook_does_not_remove_foreign_hook(
     hook_path.parent.mkdir(parents=True, exist_ok=True)
     hook_path.write_text(_FOREIGN_HOOK_CONTENT)
     monkeypatch.setattr("phi_scan.cli._shared._GIT_DIR_PATH", tmp_path / ".git")
-    monkeypatch.setattr("phi_scan.cli._PRE_COMMIT_HOOK_PATH", str(hook_path))
+    monkeypatch.setattr("phi_scan.cli.hooks._PRE_COMMIT_HOOK_PATH", str(hook_path))
 
     result = _runner.invoke(app, ["uninstall-hook"])
 
@@ -400,7 +400,7 @@ def test_install_hook_exits_with_error_when_not_in_git_repo(
 ) -> None:
     hook_path = tmp_path / ".git" / "hooks" / "pre-commit"
     monkeypatch.setattr("phi_scan.cli._shared._GIT_DIR_PATH", tmp_path / ".git")
-    monkeypatch.setattr("phi_scan.cli._PRE_COMMIT_HOOK_PATH", str(hook_path))
+    monkeypatch.setattr("phi_scan.cli.hooks._PRE_COMMIT_HOOK_PATH", str(hook_path))
 
     result = _runner.invoke(app, ["install-hook"])
 
@@ -417,7 +417,7 @@ def test_uninstall_hook_exits_with_error_when_not_in_git_repo(
 ) -> None:
     hook_path = tmp_path / ".git" / "hooks" / "pre-commit"
     monkeypatch.setattr("phi_scan.cli._shared._GIT_DIR_PATH", tmp_path / ".git")
-    monkeypatch.setattr("phi_scan.cli._PRE_COMMIT_HOOK_PATH", str(hook_path))
+    monkeypatch.setattr("phi_scan.cli.hooks._PRE_COMMIT_HOOK_PATH", str(hook_path))
 
     result = _runner.invoke(app, ["uninstall-hook"])
